% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solution.R
\name{getAssignments}
\alias{getAssignments}
\title{Get assignments}
\usage{
getAssignments(problem, solution)
}
\arguments{
\item{problem}{Problem whose model was solved.}

\item{solution}{Result of model solving (e.g. result of
\code{\link{findRepresentativeFunction}} or \code{\link{investigateUtility}}).}
}
\value{
Vector of alternative assignments. Each element contains an index
of a class that corresponding alternative was assigned to.
}
\description{
This function returns assignments for given model solution.
}
\details{
Function is deprecated. Solution already contains assignments.
}

