% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solution.R
\name{getMarginalUtilities}
\alias{getMarginalUtilities}
\title{Get marginal utilities}
\usage{
getMarginalUtilities(problem, solution)
}
\arguments{
\item{problem}{Problem whose model was solved.}

\item{solution}{Result of model solving (e.g. result of
\code{\link{findRepresentativeFunction}} or \code{\link{investigateUtility}}).}
}
\value{
A \emph{n} x \emph{m} matrix containing marginal values of \code{n} alternatives
on \code{m} criteria.
}
\description{
This function extracts alternatives marginal values from model solution.
}
\details{
Function is deprecated. Solution already contains marginal utilities.
}

