% Generated by roxygen2 (4.0.1): do not edit by hand
\name{removeMaximalClassCardinalities}
\alias{removeMaximalClassCardinalities}
\title{Remove maximal class cardinality restrictions}
\usage{
removeMaximalClassCardinalities(problem, ...)
}
\arguments{
\item{problem}{Problem from which preference information will be removed.}

\item{...}{Two-element vectors and/or integers.
Each argument represents restriction to remove. If  \code{c(i, j)} vector was
provided then defined maximal cardinality \emph{j} for class \emph{C_i} will
be removed. In case where single value \code{i} was given, a restriction for
class \emph{a_i} will be removed regardless of maximal cardinality value.
If a specific restriction was not found nothing will happen.}
}
\value{
Problem with removed preference information.
}
\description{
This function allows to remove defined maximal cardinality of particular classes.
}
\examples{
# 4 alternatives, 2 gain criteria, 3 classes, monotonously increasing
# and general marginal value functions
perf <- matrix(c(5, 2, 1, 7, 0.5, 0.9, 0.4, 0.4), ncol = 2)
problem <- buildProblem(perf, 3, FALSE, c('g', 'g'), c(0, 0))

# set maximal class cardinalities:
# at most two alternatives could be assigned to class 2
# and at most one alternative could be assigned to class 3
problem <- addMaximalClassCardinalities(problem, c(2, 2), c(3, 1))
# remove defined restriction for class 2
problem <- removeMaximalClassCardinalities(problem, 2)
}

