\name{rorutadis-package}
\alias{rorutadis-package}
\alias{rorutadis}
\docType{package}
\title{
Robust Ordinal Regression UTADIS
}
\description{
Implementation of Robust Ordinal Regression for multiple criteria value-based
sorting with some extensions and additional tools.
}
\details{
\tabular{ll}{
Package: \tab rorutadis\cr
Type: \tab Package\cr
Version: \tab 0.4.2\cr
Date: \tab 2017-01-18\cr
License: \tab GPL-3\cr
}
}
\author{
Krzysztof Ciomek

Maintainer: Krzysztof Ciomek <k.ciomek at gmail.com>
}
\keyword{ mcda ordinal package regression robust ror rorutadis sorting uta utadis }