% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solution.R
\name{getCharacteristicPoints}
\alias{getCharacteristicPoints}
\title{Get characteristic points}
\usage{
getCharacteristicPoints(problem, solution)
}
\arguments{
\item{problem}{Problem whose model was solved.}

\item{solution}{Result of model solving (e.g. result of
\code{\link{findRepresentativeFunction}} or \code{\link{investigateUtility}}).}
}
\value{
List of \code{m} matrices for each of \code{m} criteria.
Each row \code{c(g, u)} of each matrix contains coordinates of a single
characteristic point, where \code{g} - evaluation on corresponding criterion,
\code{u} - marginal utility.
}
\description{
This function extracts values of characteristic points from model solution.
}
\details{
Function is deprecated. Solution already contains characteristic points.
}

