% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/solution.R
\name{getThresholds}
\alias{getThresholds}
\title{Get thresholds}
\usage{
getThresholds(problem, solution)
}
\arguments{
\item{problem}{Problem whose model was solved.}

\item{solution}{Result of model solving (e.g. result of
\code{\link{findRepresentativeFunction}} or \code{\link{investigateUtility}}).}
}
\value{
Vector containing \code{h-1} thresholds from \code{t_1} to \code{t_h-1}
where \code{t_p-1} is lower threshold of class \code{C_p} and \code{h} is
number of classes.
}
\description{
This function extracts values of thresholds from solution.
}
\details{
Function is deprecated. Solution already contains thresholds.
}

