\name{rorutadis-package}
\alias{rorutadis-package}
\alias{rorutadis}
\docType{package}
\title{
Robust Ordinal Regression UTADIS
}
\description{
Implementation of Robust Ordinal Regression for
value-based sorting with some extensions and additional tools. It is a
novel Multiple-Criteria Decision Aiding (MCDA) framework.
}
\details{
\tabular{ll}{
Package: \tab rorutadis\cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2015-10-21\cr
License: \tab GPL-3\cr
}
}
\author{
Krzysztof Ciomek

Maintainer: Krzysztof Ciomek <k.ciomek at gmail.com>
}
\keyword{ mdca ordinal package regression robust ror rorutadis sorting uta utadis }